requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.3.0
boto3==1.26.116
botocore==1.29.116
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.7
crispy-bootstrap5==0.7
cryptography==41.0.1
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.5
django-allauth==0.54.0
django-cleanup==7.0.0
django-countries==7.5.1
django-crispy-forms==1.14.0
django-debug-toolbar==4.1.0
django-js-asset==2.0.0
django-model-utils==4.3.1
django-mptt==0.14.0
django-phonenumber-field==7.0.2
django-storages==1.13.2
exceptiongroup==1.1.0
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
jmespath==1.0.1
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
phonenumberslite==8.13.4
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
# psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
style==1.1.0
swapper==1.3.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2022.7
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.3.0
